backend/requirements.txt
fastapi
uvicorn[standard]
openai
supabase
httpx
python-dotenv
pytest
pytest-asyncio
